auth-service/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic[email]
passlib[bcrypt]
PyJWT
python-jose
python-jose[cryptography]
python-dotenv
prometheus-client
